+++ CHANGES.rst
@@ -11,6 +11,9 @@
 - Added new specutils interface
 - Calibration group numbers can now be anything, as long as there are no more
   than 63 unique integers.
+- Removed use of the term "master", renamed to calibration frames/files.
+- Calibration frame naming now done via ``calibframe.CalibFrame`` class.
+- Start to deprecate use of ``os.path`` in favor of ``pathlib``
 
 1.12.1 (21 Feb 2023)
 --------------------